Relocate some events
